refactor: reorganize repository imports for consistency
