+++ src/adapter/repository/FileSportsHallRepositoryImpl.ts
