+++ src/adapter/repository/FileTeamLeadRepositoryImpl.ts
